move igv funcs to correct file
