+++ seqr/views/apis/igv_api.py
@@ -1,3 +1,5 @@
+from collections import defaultdict
+import json
 import re
 from django.http import StreamingHttpResponse
 from django.contrib.auth.decorators import login_required
@@ -5,6 +7,8 @@ from django.contrib.auth.decorators import login_required
 from seqr.models import Individual, IgvSample
 from seqr.views.utils.dataset_utils import validate_alignment_dataset_path # TODO wrong file?
 from seqr.utils.file_utils import file_iter, does_file_exist
+from seqr.views.utils.file_utils import save_uploaded_file
+from seqr.views.utils.json_to_orm_utils import get_or_create_model_from_json
 from seqr.views.utils.json_utils import create_json_response
 from seqr.views.utils.orm_to_json_utils import  get_json_for_sample
 from seqr.views.utils.permissions_utils import get_project_and_check_permissions, check_project_permissions, \
